refactor(modules): extract MCP tools + ACP server into src/modules/protocols/
Group the external protocol-adapter surfaces under one modules/protocols/
home, mirroring the economizer relocation (#1485/#1488):

- modules/protocols/mcp/  <- mcp_tools.c, mcp_tool_profile.c,
  mcp_tools_extended.c, mcp_skill_tools.c, mcp_tools_gateway.c + headers.
  The aimee-as-MCP-server tool implementations.
- modules/protocols/acp/  <- acp_server.c + acp_server.h.
  The aimee-as-ACP-server (Agent Client Protocol, for editors like Zed).

Pure relocate: git mv + build-path updates only, no logic change. The
moved .c files had `#include "headers/mcp_*.h"` (explicit-prefix) includes;
rewrote those to bare includes resolved via the new -Imodules/protocols/mcp
path (+ -Iheaders for the stay-behind plugin.h / session_search_tool.h /
mcp_client_registry.h). Both build systems updated: Makefile C_FLAGS + the
two source lists; tests/Rules.mk object paths; CMake source refs + all six
target_include_directories. Server dispatch (server/server_mcp_*.c), the
MCP client (server/mcp_client*.c), and the git MCP tools (mcp_git_*, coupled
with git_verify) stay put for follow-up slices.

Server links; unit-test-mcp-git/-client/-gateway-tools + unit-test-acp-server
pass; make build-integrity + make lint green.
