+++ src/modules/protocols/mcp/mcp_skill_tools.h
