+++ src/modules/protocols/mcp/mcp_tools_gateway.h
